fix(core): fix format specifiers in writer.cc for cross-platform cross-compile

+++ src/core/sidecar/writer.cc
@@ -82,7 +82,7 @@ static void json_append_str(char* buf, size_t* pos, size_t cap, const char* s) {
 
 static void json_append_int(char* buf, size_t* pos, size_t cap, int64_t v) {
   char tmp[32];
-  int len = snprintf(tmp, sizeof(tmp), "%ld", v);
+  int len = snprintf(tmp, sizeof(tmp), "%lld", (long long)v);
   if (*pos + len >= cap)
     return;
   memcpy(buf + *pos, tmp, len);
@@ -493,8 +493,8 @@ pc_status pc_sidecar_try_lock(const char* doc_path) {
 
   // Write PID and timestamp
   char lock_content[128];
-  int len =
-      snprintf(lock_content, sizeof(lock_content), "pid=%d time=%ld\n", getpid(), time(nullptr));
+  int len = snprintf(lock_content, sizeof(lock_content), "pid=%d time=%lld\n", getpid(),
+                     (long long)time(nullptr));
   write_all(fd, lock_content, len);
   fsync(fd);
   close(fd);
